rename gpackage googleGamePlayServices to googlePlayGameServices. i.e. fix typo
