+++ android/src/car/superfun/game/gameModes/gladiatorMode/GladiatorMode.java
@@ -10,7 +10,7 @@ import com.badlogic.gdx.physics.box2d.FixtureDef;
 import com.badlogic.gdx.utils.Array;
 
 import car.superfun.game.GlobalVariables;
-import car.superfun.game.googleGamePlayServices.GoogleGameServices;
+import car.superfun.game.googlePlayGameServices.GoogleGameServices;
 import car.superfun.game.maps.TrackBuilder;
 import car.superfun.game.cars.LocalCarController;
 import car.superfun.game.cars.OpponentCar;
