+++ android/src/car/superfun/game/googlePlayGameServices/SignIn.java
@@ -1,4 +1,4 @@
-package car.superfun.game.googleGamePlayServices;
+package car.superfun.game.googlePlayGameServices;
 
 import android.content.Intent;
 import android.support.annotation.NonNull;
