+++ android/src/car/superfun/game/menus/GameSettings.java
@@ -12,7 +12,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import car.superfun.game.googleGamePlayServices.GoogleGameServices;
+import car.superfun.game.googlePlayGameServices.GoogleGameServices;
 import car.superfun.game.states.NewState;
 import car.superfun.game.actors.ButtonActor;
 import car.superfun.game.gameModes.gladiatorMode.GladiatorMode;
